+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## 0.3.1
+  * Add `merge` keywork
   * Format can be passed as a string or a symbol
   * Avoid to unexpectedly change cached templates (johnbintz)
   * Add full template stack support to `glue` (fnordfish)