+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+## 0.3.2 (unreleased)
+
 ## 0.3.1
   * Add `merge` keywork
   * Format can be passed as a string or a symbol